Refactored tags endpoint.
