+++ voicesofyouth/api/v1/serializers.py
@@ -1,8 +1,5 @@
-from django.shortcuts import reverse
 from rest_framework import serializers
-# from voicesofyouth.maps.models import Map
 
-from voicesofyouth.tag.models import Tag
 from voicesofyouth.users.models import User
 
 
@@ -20,11 +17,3 @@ class UserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = ('id', 'display_name', 'language', 'user_image', 'personal_url')
-
-
-class TagSerializer(serializers.ModelSerializer):
-    tag = serializers.StringRelatedField()
-
-    class Meta:
-        model = Tag
-        fields = ('id', 'tag', 'urgency_score')
